Moves dump_values, load_values, create_item, and create_section from Config to Section
